Logger Update, added system_config file
