+++ DAQ_FW/platformio.ini
@@ -13,3 +13,4 @@ platform = espressif32
 board = esp32dev
 framework = arduino
 lib_deps = thijse/ArduinoLog@^1.1.1
+build_flags = -I include
